+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added: Browser utility `CConnoisseur.closeNotes`.
+
 ## 0.2.1 - 2021-11-27
 - Added: Browser utilities `CConnoisseur.ascend` and `CConnoisseur.reincarnate`.
 - Added: Browser utility `CConnoisseur.redrawMarketMinigame`.